[FEAT] 리뷰 게시물 기능

+++ src/main/java/com/hamss2/KINO/api/admin/dto/AdminReportCommentDetailResDto.java
@@ -1,19 +1,19 @@
 package com.hamss2.KINO.api.admin.dto;
 
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class AdminReportCommentDetailResDto {
+
     private Long reportedId;
     private Long reportId;
     private Long relatedId;
-    private int relatedType;
+    private Long relatedType;
     private int reportType;
     private String reporterEmail;
     private String reportedEmail;
